chore!: use curl/tar as default on Windows (git was default before)
We changed the default method for Windows to git as older versions
of Windows had problems with the installed tar version and also with
symlinks in the repositories. This was on October 1st 2021. We expect
that users use Windows updates to have a up-to-date version installed.

We should still document how to enable the git fallback for everyone
having problems with tar in our Windows wiki. Normal users shouldn't
suffer from the slowness of downloading the whole parser tree with
generated parser.c revisions.